pyproject.toml
[project]
name = "learning-4-6-week"
version = "1.0.0"
description = "Start learning in 4-6-week, therefore create new project"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Dang Ngoc Nhan", email = "dangngocnhan27@gmail.com" }
]
dependencies = [
    "beautifulsoup4",
    "requests",
    "dlt[postgres,mongodb]",
    "dbt-core",
    "dbt-postgres",
    "pandas",
    "python-dotenv",
    "matplotlib",
    "seaborn",
    "sqlalchemy",
    "psycopg2-binary",
    "pymongo",
    "dnspython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sources"]
